Providing gpt with example problem files

+++ helpful_functions.py
@@ -1,10 +1,19 @@
 import re
 import json
+import os
 from exceptions import AttributeErrorException, WrongPathException
 
 def save_to_file(text, filepath):
-    with open(filepath, 'w+') as file:
-        file.write(text)
+    try:
+        if os.path.isfile(filepath):
+            with open(filepath, 'w+') as file:
+                file.write(text)
+        else:
+            raise WrongPathException()
+    except WrongPathException as e:
+        for i in range(10):
+            print("WRONG PATHEXC")
+            print(e)
 
 def save_lines_to_file(lines, filepath):
     with open(filepath, 'w+') as file:
@@ -24,3 +33,32 @@ def str_to_json(text):
     else:
         return
         
+def write_file(filepath, text, mode='wl'):
+    try:
+        match mode:
+            # write
+            case 'w':
+                with open(filepath, 'w+') as file:
+                    file.write(text)
+            # writelines
+            case 'wl':
+                with open(filepath, 'w+') as file:
+                    file.writelines(text)
+    except FileNotFoundError as e:
+        print('File doesnt exist')
+          
+def read_file(filepath, mode='rl'):
+    try:
+        match mode:
+            # read
+            case 'r':
+                with open(filepath, 'r') as file:
+                    file = file.read()
+            # readlines
+            case 'rl':
+                with open(filepath, 'r') as file:
+                    file = file.readlines()
+    except FileNotFoundError as e:
+        print('File doesnt exist')
+    else:  
+        return file
